+++ CHANGELOG.md
@@ -17,7 +17,8 @@ This release changes how user-created Python code is executed. This affects Pyth
 
  - Fix sunrise/sunset calculation
  - Fix inability to use "," in Input custom options
- - Fix missing install dependency for Ruuvitag Input ([#638](https://github.com/kizniche/mycodo/issues/638))
+ - Fix install dependencies for Ruuvitag Input ([#638](https://github.com/kizniche/mycodo/issues/638))
+ - Fix reliability issue with Ruuvitag Input (crashing Mycodo daemon) ([#638](https://github.com/kizniche/mycodo/issues/638))
 
 ### Features